feat: delete button is not visible on another user's profile

+++ profile.php
@@ -26,8 +26,10 @@
             <a class="edit_profile" href="edit_profile.php"><i class="fa fa-camera" aria-hidden="true" style="color:#fff; background:transparent;"></i></a>
         <?php endif; ?>
 
-        <!-- delete user profile -->
-        <a href="delete_profile.php" class="deleteUser">Delete profile</a>
+        <?php if($_SESSION["userId"] == $_GET["user"]): ?>
+            <!-- delete user profile -->
+            <a href="delete_profile.php" class="deleteUser">Delete profile</a>
+        <?php endif; ?>
     </section>
     
 </body>
